requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
distro==1.9.0
dominate==2.9.1
filelock==3.19.1
flask==3.1.2
flask-bootstrap4==4.0.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langsmith==0.4.28
markupsafe==3.0.2
matplotlib==3.10.6
mpld3==0.5.11
mpmath==1.3.0
multidict==6.6.4
narwhals==2.5.0
networkx==3.5
numpy==2.3.3
openai==1.108.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
plotly==6.3.0
propcache==0.3.2
psutil==7.0.0
pydantic==2.11.9
pydantic-core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tk==0.1.0
tokenizers==0.22.0
torch==2.8.0
torch-cluster==1.6.3
torch-geometric==2.6.1
torch-scatter==2.1.2
torch-sparse==0.6.18
torch-spline-conv==1.2.2
tqdm==4.67.1
transformers==4.56.1
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
visitor==0.1.3
werkzeug==3.1.3
yarl==1.20.1
zstandard==0.25.0
